requirements.txt
psycopg2-binary==2.9.11
python-dotenv==1.2.1
fastapi==0.120.0
gunicorn==23.0.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.3
pydantic-settings==2.11.0
